Implement sending of events
Clicking the send event button will bring up a modal where the user can
enter the event name, data and private/public flag. The event is then
sent via ParticleJS and success/failure dumped to the terminal.

Had to change the tooltip data-trigger attribute on the send event
button from the default of 'hover focus' to just 'hover'. Otherwise, the
tooltip would stay open until the modal was dismissed. Not sure what
impact this will have on mobile, may need to find a different
workaround.

The private/public checkbox is not aligning with it's label or the
tooltip question mark icon. Needs fixing.