Implement file upload to stdin
The file is split into 255 byte chunks which are sent one every second
to the Oak's stdin to avoid overloading the rx buffer or hitting the
Particle Cloud event frequency limit. The "Send Data" button is disabled
during the upload so that the user cannot interrupt the data stream.

We might consider adding a progress bar to communicate the upload
status, but I'm not sure if typical uploads will be big enough to
warrant this.

Right now, the text input field is superfluous as a path cannot be
entered with the keyboard and it must be cleared after upload, otherwise
the on-change event will not allow sending the same file twice. Perhaps
we should omit or hide the text input and keep the button only?