Independently show/hide stdin, stdout, vars, etc.
This enables each class of data printed to the terminal (stdin, stdout,
stderr, functions, variables, received events and sent events) to be
shown or hidden independently. This is achieved for current and future
elements by modifying the display property of the corresponding CSS
classes. In addition, stdout data is displayed as a span so that
newlines sent by the Oak are honored, yet non-stdout information is
always displayed on its own line. This show/hide behavior is controlled
by toggles added to the settings modal, and these settings are saved to
localStorage like the others.

Timestamps are also implemented based on @emcniece's commit
emcniece/OakTerm@7f029ff. Again, CSS is modified to create a hanging
indent when timestamps are enabled so that lines wrap tidily. Also,
stdout data is switched to `display: 'block'` since honoring newlines
doesn't make sense with timestamps enabled.

Also adds a color class for the device rename and change events.