Show hide individual content classes in terminal (stdin, stdour, stderr, events, variables, etc.) (#57)
* Echo sent data to terminal with class text_stdin

* Independently show/hide stdin, stdout, vars, etc.

This enables each class of data printed to the terminal (stdin, stdout,
stderr, functions, variables, received events and sent events) to be
shown or hidden independently. This is achieved for current and future
elements by modifying the display property of the corresponding CSS
classes. In addition, stdout data is displayed as a span so that
newlines sent by the Oak are honored, yet non-stdout information is
always displayed on its own line. This show/hide behavior is controlled
by toggles added to the settings modal, and these settings are saved to
localStorage like the others.

Timestamps are also implemented based on @emcniece's commit
emcniece/OakTerm@7f029ff. Again, CSS is modified to create a hanging
indent when timestamps are enabled so that lines wrap tidily. Also,
stdout data is switched to `display: 'block'` since honoring newlines
doesn't make sense with timestamps enabled.

Also adds a color class for the device rename and change events.

* Fix cases where stdout should have been displayed inline

* Remove <br> insertion as it's now handled in css

* Rework hide/show of content classes in terminal

This reworked approach uses `hide_*` classes which are added/removed
from the `#content` div. Static css rules change the behavior of the
`text_*` and `timestamp` classes when a `hide_*` class is present on
an ancestor element.

* Remove remaining bits of <br> insertion