Update sidebar table on button click
Resolves #39 by retrieving the variable value, updating the cached
value and the variables table, and dumping it to the terminal.

This commit also introduces a framework for error handling via an
error handler generator function designed to be called in a .catch()
block with an appropriate error message. It returns a suitable function
to deal with that error and then return a rejected promise to block
subsequent .then() actions.

Right now, this error handler simply prints the supplied error message
plus the API call error description to the console, but it can be
extended to display a modal or print the error to the terminal.