Fix error handling issue in get_variable()
If the error handler returns a rejected promise in the middle of the
chain to block later .then() clauses, an additional error handler must
be added at the end of the chain to prevent the error being "uncaught".

In this case, it was simpler to move the catch to the end of the chain
and not return a rejected promise.