WIP to fix scrollbar issues when modal displayed
When scrollbars are set to always visible, a second scrollbar is appears
when the settings modal is opened and the window size is small enough so
that the modal doesn't fit in the window. In addition, the divs
underneath the modal shift horizontally when opened, due to the
scrollbar being added on the right.

This is a work in progress using container divs to allow the scrollbar
to always be present. This way, everything stays in the same place, and
the scrollbar is shown/hidden using the css visibility property.