Add proper error handling and display
This commit adds retrying and error handling for all ParticleJS API
calls. In general, API calls are retried three times with 1 second delay
between each. If the call still does not succeed, the error is
categorized into one of three groups (`MINOR`, `MAJOR` or `FATAL`) which
dictate the action that should be taken.

For a minor error, e.g. failure to get a variable value, the error is
logged to the JS console, but not communicated to the user. For a major
error, e.g. failure to send an event, an error message is printed to the
terminal. For a fatal error, i.e. the failure of an API call the result
of which we cannot proceed without, a modal is displayed informing the
user and giving them the option to retry (which falls back into the
calling `retry_promise()` and retries three more times) or logout. When
the modal is displayed, subsequent errors are silenced, pollers are
stopped and the subscribed event stream (if there is one) is aborted.
Upon retry, the pollers and event stream are restarted.