Resolve variable update issues (#63)
* Fix bug in update_variable()

It was using the wrong selector to find the td element containing the
variable value.

* Refresh dev info when a device comes online

A device coming online (e.g. from config mode or a reboot) is signaled
by a spark/status - online event. It makes sense to refresh the device
info and variable values at this point, rather than waiting for the next
update interval. Resolves #45.

* Stop variable queries after they are unexposed

When the devinfo call returns no variables, this isEmpty check prevented
the device_vars object from being emptied. As a result, their values
were still being queried.

* Update sidebar table on button click

Resolves #39 by retrieving the variable value, updating the cached
value and the variables table, and dumping it to the terminal.

This commit also introduces a framework for error handling via an
error handler generator function designed to be called in a .catch()
block with an appropriate error message. It returns a suitable function
to deal with that error and then return a rejected promise to block
subsequent .then() actions.

Right now, this error handler simply prints the supplied error message
plus the API call error description to the console, but it can be
extended to display a modal or print the error to the terminal.

* Update variables on device change and refresh

* Fix error handling issue in get_variable()

If the error handler returns a rejected promise in the middle of the
chain to block later .then() clauses, an additional error handler must
be added at the end of the chain to prevent the error being "uncaught".

In this case, it was simpler to move the catch to the end of the chain
and not return a rejected promise.