Rework login to use new error handling
This updates the login code to use promise retries and the new error
handler. A new error category (`consts.ERR_LOGIN`) is defined for login
errors as they need to be handled differently and presented to the user
on the login screen.

This commit also adjusts the initial startup to only run `do_login()` if
we have an access token. Otherwise, the login screen is displayed.

It also fixes a logic error in `login()` where the presence of an access
token took precedence over the entered email and password upon clicking
the login button.