Streamlined retrying and error handling
Async function retrying is now moved to `Promise.prototype.retry()`.
This is a generic retry function with no OakTerm specific code. It
accepts a conditional retry handler, through which subsequent retries
can be cancelled or added in response to the particular error, and this
handler can return a promise to defer the decision.

Fatal error handling (i.e. pausing everything and displaying an error
dialog) moves into this conditional retry handler, so that it can add
a retry when the retry button is clicked. This handler also enables
skipping of retries for definitive errors that won't be fixed by
retries, i.e. invalid access token.

This relocation of the fatal error handling streamlines the promise
chains as the error handler function can move to the end of each chain,
doing away with the `catch_at_end` function.

Also added handling for unknown errors, which prints the error object
to the terminal and asks the user to report a bug.