Fix sidebar issues
Removed sidebar hover js that was breaking sidebar select dropdowns in
Firefox. Resolves #66.

Added "left: 0" to the css for fixed-sidebar, which keeps its scrollbar
onscreen when it has 100% width (previously, the sidebar was indented
slightly which pushed its scrollbar off screen). Also switched overflow
to hidden on the html element in this situation via a `sidebar-open`
class, otherwise the main content scrollbar is shown too. Resolves #72.

Changed the way the height of the fixed sidebar is defined so that it
spans only from the bottom of the header to the top of the footer.
Previously, the sidebar scrollbar would extend underneath the header and
footer to the edges of the window.