Add OakTerm logo to login and terminal screens
Two versions of the logo are included, with and without the tree, as it
doesn't render well at small scales. The tree version is used on the
login screen and the other in the upper left corner of the main terminal
screen.

Adding the main terminal logo required adjusting the column widths so
the buttons did not wrap to a second line and forcing all the buttons to
wrap below the logo at <= 570px so they did not overlap the logo.