Add proper error handling and display (#70)
This commit adds configurable retrying and error handling for all
ParticleJS API calls. In general, API calls are retried three times with
1 second delay between each. If the call still does not succeed, the
error is categorized into one of four groups (`LOGIN`, `MINOR`, `MAJOR`
or `FATAL`) which dictate the action that should be taken.

For a minor error, e.g. failure to get a variable value, the error is
logged to the JS console, but not communicated to the user. For a major
error, e.g. failure to send an event, an error message is printed to the
terminal. For a fatal error, i.e. the failure of an API call the result
of which we cannot proceed without, a modal is displayed informing the
user and giving them the option to retry (which adds an additional
retry) or logout. When the modal is displayed, subsequent errors are
silenced, pollers are stopped and the subscribed event stream (if there
is one) is aborted. Upon retry, the pollers and event stream are
restarted. Login errors are categorized separately as they are reported
directly on the login screen.

Retries are implemented as `.then(retry(async_func, conditional, ...))`.
This is a generic retry function with no OakTerm specific code. It
accepts a conditional retry handler, through which subsequent retries
can be cancelled or added in response to the particular error, and this
handler can return a promise to defer the decision. Fatal error handling
(i.e. pausing everything and displaying an error dialog) takes place in
this conditional retry handler, so that it can add a retry when the
retry button is clicked. This handler also enables skipping of retries
for definitive errors that won't be fixed by retries, i.e. invalid
access token.

Certain errors are regarded as definitive, that is, retrying the
operation will not change the outcome. These are defined in
`definitive_error()` and the conditional handler will cancel any
subequent retries if it encounters a definitive error.
 
This commit also makes the following minor changes:

* Adjust the initial startup to only run `do_login()` if we have an 
  access token. Otherwise, the login screen is displayed.
* Fix a logic error in `login()` where the presence of an access token
  took precedence over the entered email and password upon clicking the
  login button.
* Disable the login button after clicking to avoid a flood if the user
  clicks multiple times.