Add OakTerm logo to login and terminal screens (#79)
* Revert reboot dropdown to regular button

Needed for MVP as we don't have functionality in place to reboot all
devices. Also, the motivation of using a dropdown was to prevent the
user from accidentally rebooting, but for consistent behavior, we'd need
to also dropdown the config mode button as it also reboots the Oak.

* Adjust styling of show-sidebar icon

Its line-height was increasing the height of the header, and its
position was slightly different in the xs layout.

* Adjust media query breakpoints to match Bootstrap

We were off by one.

* Fix capitalization in login placeholder

* Add OakTerm logo to login and terminal screens

Two versions of the logo are included, with and without the tree, as it
doesn't render well at small scales. The tree version is used on the
login screen and the other in the upper left corner of the main terminal
screen.

Adding the main terminal logo required adjusting the column widths so
the buttons did not wrap to a second line and forcing all the buttons to
wrap below the logo at <= 570px so they did not overlap the logo.