Ensure sent data prints before subsequent responses
Due to the asynchronous nature of the Particle API calls, responses to
sent data were being printed before the sent data was printed itself.
This commit moves the `terminal_print()` call before the send to avoid
this issue.

A side effect of this change is that the sent data will be printed even
if there is an error sending. I think this is a reasonable trade-off,
given an error message will be printed informing the user that the send
failed.