Fix bug in var color coding and tweak colors
The integer css class should have been named `var-type-int32` not
`var-type-int`, so it wasn't being applied.

Also darkened the yellow coloring as the contrast to the light gray
button text was poor, and adjusted the string button coloring as it
didn't darken when hovered like the others (there was probably a mixup
between the background/border/hover colors).