Fix unhandled error when no devices are registered
When no devices are registered with the logged in Particle account, an
unhandled error would occur because the code would attempt to get device
info for an undefined current device.

This commit resolves the issue by aborting the promise chain at the
`get_devinfo()` call, as the remainder of the chain is irrelevant
without a registered device. In this case, the Rename Device button
under Device Details in the sidebar also needs to be disabled, as
clicking it would generate an error.